# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/tickets": {
      "post": {
        "operationId": "TicketsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Tickets"
        ]
      },
      "get": {
        "operationId": "TicketsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TicketResponse"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Tickets"
        ]
      }
    },
    "/v1/tickets/sort": {
      "post": {
        "operationId": "TicketsController_sort",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SortColumnDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Tickets"
        ]
      }
    },
    "/v1/tickets/{ticketId}": {
      "get": {
        "operationId": "TicketsController_get",
        "parameters": [
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Tickets"
        ]
      },
      "patch": {
        "operationId": "TicketsController_update",
        "parameters": [
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Tickets"
        ]
      },
      "delete": {
        "operationId": "TicketsController_remove",
        "parameters": [
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Tickets"
        ]
      }
    },
    "/v1/tickets/{ticketId}/position": {
      "put": {
        "operationId": "TicketsController_move",
        "parameters": [
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovePositionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Tickets"
        ]
      }
    }
  },
  "info": {
    "title": "To-do API",
    "description": "티켓 보드 API. 정본은 docs/api_spec.md.",
    "version": "1",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CreateTicketDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "description": "앞뒤 공백은 제거된다."
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 2000
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ],
            "default": "MEDIUM"
          },
          "dueAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "tags": {
            "description": "앞뒤 공백 제거·소문자로 정규화, 중복 제거. 최대 10개, 이름 최대 30자.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title"
        ]
      },
      "TicketResponse": {
        "type": "object",
        "properties": {
          "ticketId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "TODO",
              "IN_PROGRESS",
              "DONE"
            ]
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ]
          },
          "dueAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "tags": {
            "description": "이름 오름차순.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "ticketId",
          "title",
          "description",
          "status",
          "priority",
          "dueAt",
          "tags",
          "createdAt",
          "updatedAt"
        ]
      },
      "SortColumnDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "TODO",
              "IN_PROGRESS",
              "DONE"
            ]
          },
          "sortBy": {
            "type": "string",
            "enum": [
              "PRIORITY",
              "DUE_AT"
            ]
          },
          "direction": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ]
          }
        },
        "required": [
          "status",
          "sortBy",
          "direction"
        ]
      },
      "UpdateTicketDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "description": "앞뒤 공백은 제거된다."
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 2000,
            "description": "null·빈 문자열·공백뿐이면 설명이 지워진다."
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ]
          },
          "dueAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "null이면 마감일이 지워진다."
          },
          "tags": {
            "description": "앞뒤 공백 제거·소문자로 정규화, 중복 제거. 최대 10개, 이름 최대 30자.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "MovePositionDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "TODO",
              "IN_PROGRESS",
              "DONE"
            ],
            "description": "이동할 대상 컬럼."
          },
          "anchorTicketId": {
            "type": "string",
            "format": "uuid",
            "description": "기준 카드. 대상 컬럼에 다른 카드가 없으면 생략한다."
          },
          "placement": {
            "type": "string",
            "enum": [
              "BEFORE",
              "AFTER"
            ],
            "description": "기준 카드의 앞 또는 뒤."
          }
        },
        "required": [
          "status",
          "placement"
        ]
      }
    }
  }
}